+++ src/chess/pieces/Pawn.java
@@ -5,7 +5,7 @@ import java.util.*;
 
 public class Pawn extends Piece {
     public static final String NOTATION = "";
-    public static final double VALUE = 1;
+    public static final int VALUE = 1;
 
     public Pawn(Board board, Square position, Color color) {
         super(board, position, color, VALUE, NOTATION);
